+++ tests/libs/fixtures/__init__.py
@@ -1,6 +1,7 @@
 from .cli import default_cli_args
 from .network import free_port
 from .web3 import contracts_manager, keystore_file, mockchain, wait_for_blocks
+from .metrics import prometheus_client_teardown, prometheus_client_collectors
 
 # Without declaring `__all__`, the names `cli` and `web3` would get imported
 # when doing `from libs.fixtures import *`, which is not intended.
@@ -11,4 +12,6 @@ __all__ = [
     "mockchain",
     "default_cli_args",
     "free_port",
+    "prometheus_client_collectors",
+    "prometheus_client_teardown"
 ]
