+++ tests/monitoring/fixtures/__init__.py
@@ -2,3 +2,4 @@ from .api import *  # noqa
 from .contracts import *  # noqa
 from .factories import *  # noqa
 from .server import *  # noqa
+from .metrics import *  # noqa
